+++ src/server/rendering/plainRendering.js
@@ -1,0 +1,5 @@
+import template from './template';
+
+export default (req, res) => res
+  .status(200)
+  .send(template());
